Replaced the KeyGenWorkers table with a 'job' field on a profile, updated all DB commands to match the change, and added a 'characters' folder to manage characters, including the 'job' enum that defines what work a character is assigned to.  Validation of econ/profile script changes to come.
