+++ src/characters/RarityClass.py
