+++ src/db/MariadbIfc.py
@@ -15,9 +15,10 @@ from mariadb.constants import *
 import os
 import pathlib as pl
 import pickle as pic
-import src.utilities.ProfileGenerator as pg
-import src.utilities.RarityClass as rc
-import src.utilities.StatsClass as sc
+import src.characters.CharacterJobs as cj
+import src.characters.ProfileGenerator as pg
+import src.characters.RarityClass as rc
+import src.characters.StatsClass as sc
 import sys
 import threading as th
 from typing import Literal, Optional
@@ -98,19 +99,16 @@ class MariadbIfc:
                 paths = { 'db'   : pl.Path('src/db/db_commands.json').absolute(),
                           'econ' : pl.Path('src/db/queries/economy_queries.json').absolute(),
                           'inv'  : pl.Path('src/db/queries/inventory_queries.json').absolute(),
-                          'keyg' : pl.Path('src/db/queries/keygen_queries.json').absolute(),
                           'pic'  : pl.Path('src/db/queries/picture_queries.json').absolute(),
                           'prof' : pl.Path('src/db/queries/profile_queries.json').absolute(),
                           'user' : pl.Path('src/db/queries/user_queries.json').absolute()}
-                self.db_log.debug(f"paths are: {paths['db']} {paths['econ']} {paths['inv']} {paths['keyg']} {paths['pic']} {paths['prof']} {paths['user']}")
+                self.db_log.debug(f"paths are: {paths['db']} {paths['econ']} {paths['inv']} {paths['pic']} {paths['prof']} {paths['user']}")
                 json_file         = open(paths['db'])
                 self.db_cmds      = json.load(json_file)
                 json_file         = open(paths['econ'])
                 self.cmds['econ'] = json.load(json_file)
                 json_file         = open(paths['inv'])
                 self.cmds['inv']  = json.load(json_file)
-                json_file         = open(paths['keyg'])
-                self.cmds['keyg'] = json.load(json_file)
                 json_file         = open(paths['pic'])
                 self.cmds['pic']  = json.load(json_file)
                 json_file         = open(paths['prof'])
@@ -131,7 +129,7 @@ class MariadbIfc:
                 raise PermissionError(f"Unable to access mariaDB server!")
 
             self.db_log.info(f"Successfully connected to database: host: {options['host']} port: {options['port']} username: {options['user_name']} db: {options['database']}")
-            self.db_log.debug(f"Loaded commands: {self.db_cmds} {self.cmds['pic']} {self.cmds['econ']} {paths['inv']} {self.cmds['keyg']} {self.cmds['prof']} {self.cmds['user']}")
+            self.db_log.debug(f"Loaded commands: {self.db_cmds} {self.cmds['pic']} {self.cmds['econ']} {paths['inv']} {self.cmds['prof']} {self.cmds['user']}")
 
     def assignKeyGenWork(self,
                          count       : int,
@@ -153,35 +151,33 @@ class MariadbIfc:
         """
         cmd        = ""
         cursor     = self.con.cursor(buffered=False)
-        key        = f'put_tier_{tier}'
+        work_tier  = cj.CharacterJobTypeEnum.KEY_GENERATION_t0.value + tier
         #This is a workaround to the cursor interpreting None as 'None'
-        new_entry  = [INDICATOR.NULL,INDICATOR.NULL,INDICATOR.NULL,INDICATOR.NULL,INDICATOR.NULL, user_id]
+        new_entry  = [work_tier, user_id, INDICATOR.NULL,INDICATOR.NULL,INDICATOR.NULL,INDICATOR.NULL,INDICATOR.NULL]
         worker_cnt = count
 
         #The remove function ensures that any existing workers always occupy
         #the front slots (and that empty slots are at the end)
         for worker in range(0, tier_data['count']):
 
-            new_entry[worker] = tier_data['workers'][worker]
+            #offset for the front-loaded spaces in new_entry
+            new_entry[worker + 2] = tier_data['workers'][worker]
             worker_cnt += 1
 
         for slot in range(0, len(profile_ids)) :
 
-            new_entry[slot + tier_data['count']] = profile_ids[slot]
+            #offset for the front-loaded spaces in new_entry
+            new_entry[slot + tier_data['count'] + 2] = profile_ids[slot]
             worker_cnt += 1
-
-            cmd = (self.cmds['prof']['put_occupied']) % (True, profile_ids[slot])
-            self.db_log.debug(f"Preparing to mark a profile as occupied: {cmd}")
-            cursor.execute(cmd)
 
         data = tuple(new_entry)
 
-        cmd  = (self.cmds['keyg'][key]) % (new_entry[0], new_entry[1], new_entry[2], new_entry[3], new_entry[4], user_id)
+        cmd  = (self.cmds['prof']['put_workers']) % (work_tier, user_id, new_entry[0], new_entry[1], new_entry[2], new_entry[3], new_entry[4])
         self.db_log.debug(f"Updating user's keygen work list for tier {tier}: {cmd}")
         #This must be done as implemented due to how the mariadb python cursor
         #handles (or rather, doesn't handle) NULL entries. See
         #https://mariadb-corporation.github.io/mariadb-connector-python/usage.html#using-indicators
-        cursor.execute(self.cmds['keyg'][key], data)
+        cursor.execute(self.cmds['prof']['put_workers'], data)
 
         cmd = (self.cmds['econ']['put_keygen_count']) % (worker_cnt, user_id)
         self.db_log.debug(f"Updating user's econ keygen count: {cmd}")
@@ -223,11 +219,6 @@ class MariadbIfc:
             self.db_log.debug(f"Creating user {id}'s ivnentory table: {cmd}")
             cursor.execute(cmd)
             self.db_log.info(f"Creatied user {id}'s inventory table")
-
-            cmd = (self.cmds['keyg']['put_new']) % (id)
-            self.db_log.debug(f"Preparing to create user keygen entry: {cmd}")
-            cursor.execute(cmd)
-            self.db_log.info(f"Updated user's keygen worker entries.")
 
             result = True
 
@@ -294,12 +285,12 @@ class MariadbIfc:
 
         prof = {}
 
-        cmd = (self.cmds['keyg']['get_tier_all']) % (user_id)
+        cmd = (self.cmds['prof']['get_workers']) % (cj.CharacterJobTypeEnum.KEY_GENERATION_t0.value, cj.CharacterJobTypeEnum.KEY_GENERATION_t5.value, user_id)
         self.db_log.debug(f"Getting current user keygen worker allocation: {cmd}")
         cursor.execute(cmd)
         #This is returned as a tuple, so either way it has to be converted to a
         #dict at some point.
-        result = cursor.fetchone()
+        result = cursor.fetchall()
         self.db_log.debug(f"Current user keygen allocation is: {result}")
         prof['workers'] = self.mapQueryToKeyGenInfo(query=result)
         self.db_log.debug(f"Worker stats are: {prof['workers']}")
@@ -479,12 +470,12 @@ class MariadbIfc:
 
         prof = {}
 
-        cmd = (self.cmds['keyg']['get_tier_all']) % (user_id)
+        cmd = (self.cmds['prof']['get_workers']) % (cj.CharacterJobTypeEnum.KEY_GENERATION_t0.value, cj.CharacterJobTypeEnum.KEY_GENERATION_t5.value ,user_id)
         self.db_log.debug(f"Getting current user keygen worker allocation: {cmd}")
         cursor.execute(cmd)
         #This is returned as a tuple, so either way it has to be converted to a
         #dict at some point.
-        result = cursor.fetchone()
+        result = cursor.fetchall()
         self.db_log.debug(f"Current user keygen allocation is: {result}")
         prof['workers'] = self.mapQueryToKeyGenInfo(query=result)
         self.db_log.debug(f"Worker stats are: {prof['workers']}")
@@ -507,7 +498,7 @@ class MariadbIfc:
         cmd     = ""
         cursor  = self.con.cursor(buffered=False)
         #This is a workaround to the cursor interpreting None as 'None'.
-        ids    = [INDICATOR.NULL for x in range(0, int(self.cmds['keyg']['max_workers']))]
+        ids    = [INDICATOR.NULL for x in range(0, int(self.cmds['econ']['max_workers_per_tier']))]
         index   = 0
         results = []
 
@@ -686,7 +677,7 @@ class MariadbIfc:
         result = cursor.fetchone()
 
         if result:
-        
+
             results['dust'] = result[1]
 
             for tier in range(0,6):
@@ -738,18 +729,26 @@ class MariadbIfc:
 
            Output: dict - a dict of the results grouped by tier.
         """
-        workers = {'tier_0' : {'workers' : [query[ 0], query[ 1],query[ 2],query[ 3],query[ 4]]},
-                   'tier_1' : {'workers' : [query[ 5], query[ 6],query[ 7],query[ 8],query[ 9]]},
-                   'tier_2' : {'workers' : [query[10], query[11],query[12],query[13],query[14]]},
-                   'tier_3' : {'workers' : [query[15], query[16],query[27],query[28],query[29]]},
-                   'tier_4' : {'workers' : [query[20], query[21],query[22],query[23],query[24]]},
-                   'tier_5' : {'workers' : [query[25], query[26],query[27],query[28],query[29]]}}
-        workers['tier_0']['count'] = len(workers['tier_0']['workers']) - workers['tier_0']['workers'].count(None)
-        workers['tier_1']['count'] = len(workers['tier_1']['workers']) - workers['tier_1']['workers'].count(None)
-        workers['tier_2']['count'] = len(workers['tier_2']['workers']) - workers['tier_2']['workers'].count(None)
-        workers['tier_3']['count'] = len(workers['tier_3']['workers']) - workers['tier_3']['workers'].count(None)
-        workers['tier_4']['count'] = len(workers['tier_4']['workers']) - workers['tier_4']['workers'].count(None)
-        workers['tier_5']['count'] = len(workers['tier_5']['workers']) - workers['tier_5']['workers'].count(None)
+        ID      = 0
+        JOB     = 1
+        workers = {'tier_0' : {'workers' : []},
+                   'tier_1' : {'workers' : []},
+                   'tier_2' : {'workers' : []},
+                   'tier_3' : {'workers' : []},
+                   'tier_4' : {'workers' : []},
+                   'tier_5' : {'workers' : []}}
+
+        for info in query:
+
+            tier = int(info[JOB]) - cj.CharacterJobTypeEnum.KEY_GENERATION_t0.value
+            workers[f'tier_{tier}']['workers'] = info[ID]
+
+        workers['tier_0']['count'] = len(workers['tier_0']['workers'])
+        workers['tier_1']['count'] = len(workers['tier_1']['workers'])
+        workers['tier_2']['count'] = len(workers['tier_2']['workers'])
+        workers['tier_3']['count'] = len(workers['tier_3']['workers'])
+        workers['tier_4']['count'] = len(workers['tier_4']['workers'])
+        workers['tier_5']['count'] = len(workers['tier_5']['workers'])
 
         return workers
 
@@ -858,24 +857,16 @@ class MariadbIfc:
 
         data = tuple(new_entry)
 
-        cmd  = (self.cmds['keyg'][key]) % (new_entry[0], new_entry[1], new_entry[2], new_entry[3], new_entry[4], user_id)
+        cmd  = (self.cmds['prof']['put_workers']) % (work_tier, user_id, new_entry[0], new_entry[1], new_entry[2], new_entry[3], new_entry[4])
         self.db_log.debug(f"Updating user's keygen work list for tier {tier}: {cmd}")
         #This must be done as implemented due to how the mariadb python cursor
         #handles (or rather, doesn't handle) NULL entries. See
         #https://mariadb-corporation.github.io/mariadb-connector-python/usage.html#using-indicators
-        cursor.execute(self.cmds['keyg'][key], data)
+        cursor.execute(self.cmds['prof']['put_workers'], data)
 
         cmd = (self.cmds['econ']['put_keygen_count']) % (count - len(profile_ids), user_id)
         self.db_log.debug(f"Updating user's econ keygen count: {cmd}")
         cursor.execute(cmd)
-
-        #Only update occupied after clearing the work table in case the command
-        #fails somehow, ensuring that the table won't have ghost workers.
-        for slot in range(0, len(profile_ids)) :
-
-            cmd = (self.cmds['prof']['put_occupied']) % (False, profile_ids[slot])
-            self.db_log.debug(f"Preparing to mark a profile as unoccupied: {cmd}")
-            cursor.execute(cmd)
 
     def resetDailyRoll(self):
         """Resets the 'daily' boolean for all user profiles, allowing them to
@@ -1014,7 +1005,9 @@ class MariadbIfc:
         """
         cmd    = ""
         cursor = self.con.cursor(buffered=False)
+        JOB    = 1
         result = None
+        USER   = 0
 
         self.db_log.warning(f"Preparing to update daily Key Gen counts.")
 
@@ -1023,32 +1016,42 @@ class MariadbIfc:
             #TODO: find the sordid single-line statement capable of doing this,
             #instead of by-user.  The current user list is small enough that
             #the waste is okay, but it should be corrected.
-            cmd = (self.cmds['keyg']['get_dailys'])
+            cmd = (self.cmds['prof']['get_workers_daily'])
             self.db_log.debug(f"Executing update daily Key Gen counts command: {cmd}")
             cursor.execute(cmd)
 
-            #Yes, this must be done this way because the subsequent cursor
-            #interactions will delete the query results.
             results = cursor.fetchall()
             self.db_log.debug(f"Found users to udpate: {results}")
 
+            user_id = result[0][USER]
+            counts  = {'t0' : 0,
+                       't1' : 0,
+                       't2' : 0,
+                       't3' : 0,
+                       't4' : 0,
+                       't5' : 0}
+
             #Reminder that the cursor returns all results as tuples.
-            for usr in results:
+            for row in results:
 
+                if row[USER] != user_id:
 
-                cmd = (self.cmds['keyg']['get_tier_all']) % (usr[0])
-                self.db_log.debug(f"Getting current user keygen worker allocation: {cmd}")
-                cursor.execute(cmd)
+                    cmd = (self.cmds['inv']['put_daily']) % (counts['t0'], counts['t1'], counts['t2'], counts['t3'], counts['t4'], counts['t5'], user_id)
+                    self.db_log.debug(f"Updating users keys with new values: {cmd}")
+                    cursor.execute(cmd)
+                    self.db_log.info(f"Updated user {usr[0]}'s keys")
 
-                result = cursor.fetchone()
-                self.db_log.debug(f"Current user keygen allocation is: {result}")
-                workers = self.mapQueryToKeyGenInfo(query=result)
-                self.db_log.debug(f"Worker stats are: {workers}")
+                    counts  = {'t0' : 0,
+                               't1' : 0,
+                               't2' : 0,
+                               't3' : 0,
+                               't4' : 0,
+                               't5' : 0}
 
-                cmd = (self.cmds['inv']['put_daily']) % (workers['tier_0']['count'], workers['tier_1']['count'], workers['tier_2']['count'], workers['tier_3']['count'], workers['tier_4']['count'], workers['tier_5']['count'], usr[0])
-                self.db_log.debug(f"Updating users keys with new values: {cmd}")
-                cursor.execute(cmd)
-                self.db_log.info(f"Updated user {usr[0]}'s keys")
+                else:
+
+                    tier = int(row[JOB]) - cj.CharacterJobTypeEnum.KEY_GENERATION_t0.value
+                    counts[f't{tier}'] += 1
 
         except Exception as err:
