+++ src/db/queries/inventory_queries.json
@@ -2,7 +2,7 @@
 	"del_default"   : "DELETE FROM IGSDInventory WHERE (u_ID = 1) LIMIT 1;",
 	"get_inventory" : "SELECT * FROM IGSDInventory WHERE (u_ID = %s);",
 	"make_def_tst"  : "INSERT INTO IGSDInventory VALUES (1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0);",
-	"put_daily"     : "UPDATE IGSDInventory SET t0_key_count=t0_key_count+%s, t1_key_count=t1_key_count+%s, t2_key_count=t2_key_count+%s, t3_key_count=t3_key_count+%s, t4_key_count=t4_key_count+%s, t5_key_count=t5_key_count+%s WHERE (u_ID = %s);",
+	"put_key_daily" : "UPDATE IGSDInventory SET t0_key_count=t0_key_count+%s, t1_key_count=t1_key_count+%s, t2_key_count=t2_key_count+%s, t3_key_count=t3_key_count+%s, t4_key_count=t4_key_count+%s, t5_key_count=t5_key_count+%s WHERE (u_ID = %s);",
 	"put_new"       : "INSERT IGNORE INTO IGSDInventory VALUES ('%s', 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0);",
 	"table_fmt"     : "IGSDInventory (u_ID BIGINT UNSIGNED NOT NULL UNIQUE PRIMARY KEY, dust BIGINT UNSIGNED NOT NULL DEFAULT 0, t0_armor_count  INT UNSIGNED NOT NULL DEFAULT 0, t0_key_count INT UNSIGNED NOT NULL DEFAULT 1, t0_weapon_count INT UNSIGNED NOT NULL DEFAULT 0, t1_armor_count INT UNSIGNED NOT NULL DEFAULT 0, t1_key_count INT UNSIGNED NOT NULL DEFAULT 0, t1_weapon_count INT UNSIGNED NOT NULL DEFAULT 0, t2_armor_count INT UNSIGNED NOT NULL DEFAULT 0, t2_key_count INT UNSIGNED NOT NULL DEFAULT 0, t2_weapon_count INT UNSIGNED NOT NULL DEFAULT 0, t3_armor_count INT UNSIGNED NOT NULL DEFAULT 0, t3_key_count INT UNSIGNED NOT NULL DEFAULT 0, t3_weapon_count INT UNSIGNED NOT NULL DEFAULT 0, t4_armor_count INT UNSIGNED NOT NULL DEFAULT 0, t4_key_count INT UNSIGNED NOT NULL DEFAULT 0, t4_weapon_count INT UNSIGNED NOT NULL DEFAULT 0, t5_armor_count INT UNSIGNED NOT NULL DEFAULT 0, t5_key_count INT UNSIGNED NOT NULL DEFAULT 0, t5_weapon_count INT UNSIGNED NOT NULL DEFAULT 0);"
 }
