--- src/db/queries/keygen_queries.json
@@ -1,23 +1,0 @@
-{
-	"del_default"  : "DELETE FROM IGSDKeyGenWorkers WHERE (u_ID = 1) LIMIT 1;",
-	"get_all"      : "SELECT * FROM IGSDKeyGenWorkers;",
-	"get_dailys"   : "SELECT u_ID FROM IGSDKeyGenWorkers WHERE (SELECT COALESCE(tier_0_slot_1, tier_0_slot_2, tier_0_slot_3, tier_0_slot_4, tier_0_slot_5, tier_1_slot_1, tier_1_slot_2, tier_1_slot_3, tier_1_slot_4, tier_1_slot_5, tier_2_slot_1, tier_2_slot_2, tier_2_slot_3, tier_2_slot_4, tier_2_slot_5, tier_3_slot_1, tier_3_slot_2, tier_3_slot_3, tier_3_slot_4, tier_3_slot_5,tier_4_slot_1, tier_4_slot_2, tier_4_slot_3, tier_4_slot_4, tier_4_slot_5, tier_5_slot_1, tier_5_slot_2, tier_5_slot_3, tier_5_slot_4, tier_5_slot_5)) IS NOT NULL;",
-	"get_tier_all" : "SELECT tier_0_slot_1, tier_0_slot_2, tier_0_slot_3, tier_0_slot_4, tier_0_slot_5, tier_1_slot_1, tier_1_slot_2, tier_1_slot_3, tier_1_slot_4, tier_1_slot_5, tier_2_slot_1, tier_2_slot_2, tier_2_slot_3, tier_2_slot_4, tier_2_slot_5, tier_3_slot_1, tier_3_slot_2, tier_3_slot_3, tier_3_slot_4, tier_3_slot_5,tier_4_slot_1, tier_4_slot_2, tier_4_slot_3, tier_4_slot_4, tier_4_slot_5, tier_5_slot_1, tier_5_slot_2, tier_5_slot_3, tier_5_slot_4, tier_5_slot_5 FROM IGSDKeyGenWorkers WHERE (u_ID = %s);",
-	"get_tier_0"   : "SELECT tier_0_slot_1, tier_0_slot_2, tier_0_slot_3, tier_0_slot_4, tier_0_slot_5 FROM IGSDKeyGenWorkers WHERE (u_ID = %s);",
-	"get_tier_1"   : "SELECT tier_1_slot_1, tier_1_slot_2, tier_1_slot_3, tier_1_slot_4, tier_1_slot_5 FROM IGSDKeyGenWorkers WHERE (u_ID = %s);",
-	"get_tier_2"   : "SELECT tier_2_slot_1, tier_2_slot_2, tier_2_slot_3, tier_2_slot_4, tier_2_slot_5 FROM IGSDKeyGenWorkers WHERE (u_ID = %s);",
-	"get_tier_3"   : "SELECT tier_3_slot_1, tier_3_slot_2, tier_3_slot_3, tier_3_slot_4, tier_3_slot_5 FROM IGSDKeyGenWorkers WHERE (u_ID = %s);",
-	"get_tier_4"   : "SELECT tier_4_slot_1, tier_4_slot_2, tier_4_slot_3, tier_4_slot_4, tier_4_slot_5 FROM IGSDKeyGenWorkers WHERE (u_ID = %s);",
-	"get_tier_5"   : "SELECT tier_5_slot_1, tier_5_slot_2, tier_5_slot_3, tier_5_slot_4, tier_5_slot_5 FROM IGSDKeyGenWorkers WHERE (u_ID = %s);",
-	"make_def_tst" : "INSERT INTO IGSDKeyGenWorkers VALUES (1, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL,NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL);",
-	"max_in_tier"  : "5",
-	"max_workers"  : "30",
-	"put_new"      : "INSERT IGNORE INTO IGSDKeyGenWorkers VALUES ('%s', NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL,NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL);",
-	"put_tier_0"   : "UPDATE IGSDKeyGenWorkers SET tier_0_slot_1=%s, tier_0_slot_2=%s, tier_0_slot_3=%s, tier_0_slot_4=%s, tier_0_slot_5=%s WHERE (u_ID = %s);",
-	"put_tier_1"   : "UPDATE IGSDKeyGenWorkers SET tier_1_slot_1=%s, tier_1_slot_2=%s, tier_1_slot_3=%s, tier_1_slot_4=%s, tier_1_slot_5=%s WHERE (u_ID = %s);",
-	"put_tier_2"   : "UPDATE IGSDKeyGenWorkers SET tier_2_slot_1=%s, tier_2_slot_2=%s, tier_2_slot_3=%s, tier_2_slot_4=%s, tier_2_slot_5=%s WHERE (u_ID = %s);",
-	"put_tier_3"   : "UPDATE IGSDKeyGenWorkers SET tier_3_slot_1=%s, tier_3_slot_2=%s, tier_3_slot_3=%s, tier_3_slot_4=%s, tier_3_slot_5=%s WHERE (u_ID = %s);",
-	"put_tier_4"   : "UPDATE IGSDKeyGenWorkers SET tier_4_slot_1=%s, tier_4_slot_2=%s, tier_4_slot_3=%s, tier_4_slot_4=%s, tier_4_slot_5=%s WHERE (u_ID = %s);",
-	"put_tier_5"   : "UPDATE IGSDKeyGenWorkers SET tier_5_slot_1=%s, tier_5_slot_2=%s, tier_5_slot_3=%s, tier_5_slot_4=%s, tier_5_slot_5=%s WHERE (u_ID = %s);",
-	"table_fmt"    : "IGSDKeyGenWorkers (u_ID BIGINT UNSIGNED NOT NULL UNIQUE PRIMARY KEY, tier_0_slot_1 UUID, tier_0_slot_2 UUID, tier_0_slot_3 UUID, tier_0_slot_4 UUID, tier_0_slot_5 UUID, tier_1_slot_1 UUID, tier_1_slot_2 UUID, tier_1_slot_3 UUID, tier_1_slot_4 UUID, tier_1_slot_5 UUID, tier_2_slot_1 UUID, tier_2_slot_2 UUID, tier_2_slot_3 UUID, tier_2_slot_4 UUID, tier_2_slot_5 UUID, tier_3_slot_1 UUID, tier_3_slot_2 UUID, tier_3_slot_3 UUID, tier_3_slot_4 UUID, tier_3_slot_5 UUID, tier_4_slot_1 UUID, tier_4_slot_2 UUID, tier_4_slot_3 UUID, tier_4_slot_4 UUID, tier_4_slot_5 UUID, tier_5_slot_1 UUID, tier_5_slot_2 UUID, tier_5_slot_3 UUID, tier_5_slot_4 UUID, tier_5_slot_5 UUID);"
-}
